--- tools/ai_review/weekly_review.py
@@ -1,66 +1,0 @@
-from tools.ai_review.cli import DEFAULT_OUTPUT_DIR
-from tools.ai_review.cli import main
-from tools.ai_review.cli import parse_args
-from tools.ai_review.cli import run_from_args
-from tools.ai_review.config import Limits
-from tools.ai_review.config import RuntimeConfig
-from tools.ai_review.config import Scope
-from tools.ai_review.config import load_config
-from tools.ai_review.exceptions import ReviewError
-from tools.ai_review.model import build_model_uri
-from tools.ai_review.paths import is_allowed_by_scope
-from tools.ai_review.paths import is_probably_binary
-from tools.ai_review.paths import match_glob
-from tools.ai_review.paths import posix_relative
-from tools.ai_review.paths import repo_root_from_script
-from tools.ai_review.paths import sanitize_text
-from tools.ai_review.paths import within_path
-from tools.ai_review.process_env import safe_subprocess_env
-from tools.ai_review.prompts import build_initial_messages
-from tools.ai_review.prompts import gather_top_level
-from tools.ai_review.prompts import render_focus
-from tools.ai_review.recent_activity import gather_recent_activity
-from tools.ai_review.runner import format_tool_result
-from tools.ai_review.runner import run_review
-from tools.ai_review.schema import build_json_schema
-from tools.ai_review.schema import validate_action
-from tools.ai_review.agent_tools import ReviewTools
-from tools.ai_review.yandex_client import API_URL
-from tools.ai_review.yandex_client import YandexClient
-
-
-__all__ = [
-    "API_URL",
-    "DEFAULT_OUTPUT_DIR",
-    "Limits",
-    "ReviewError",
-    "ReviewTools",
-    "RuntimeConfig",
-    "Scope",
-    "YandexClient",
-    "build_initial_messages",
-    "build_json_schema",
-    "build_model_uri",
-    "format_tool_result",
-    "gather_recent_activity",
-    "gather_top_level",
-    "is_allowed_by_scope",
-    "is_probably_binary",
-    "load_config",
-    "main",
-    "match_glob",
-    "parse_args",
-    "posix_relative",
-    "render_focus",
-    "repo_root_from_script",
-    "run_from_args",
-    "run_review",
-    "safe_subprocess_env",
-    "sanitize_text",
-    "validate_action",
-    "within_path",
-]
-
-
-if __name__ == "__main__":
-    raise SystemExit(main())
