+++ tools/warden/exceptions.py
